Patch for issue 1535 from Steve Horgan: System Information page.
This page is provides system administrators with a high-level overview of the
running Mifos instance.

svnmerge didn't do the job properly (probably because the branch started out as
a snapshot of edits to the trunk instead of a "clean" branch off the trunk) so
I resorted to a manual merge of the issue1535 branch. I manually removed the
svnmerge tracking information.

Merge command was:

svn merge \
  https://mifos.dev.java.net/svn/mifos/trunk@12529 \
  https://mifos.dev.java.net/svn/mifos/branches/issue1535


git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@12539 a8845c50-7012-0410-95d3-8e1449b9b1e4