Applying patch from Stephen Horgan to read .svn/entries directly rather than using svnant. Thanks, Steve!
Added SystemInfoTest to ApplicationTestSet4 so it is part of the full
unit test suite. Not sure if this is the right set to add it to, but we
can always shift these around later if necessary.

Updated SystemInfoTest to use JUnit 4 since, in my humble opinion, JUnit
4 is clearer and a bit more concise. Also, I believe we want to
eventually move entirely to JUnit 4.

Relevant threads:
http://thread.gmane.org/gmane.comp.finance.mifos.devel/4143/focus=4255
http://thread.gmane.org/gmane.comp.finance.mifos.devel/4275


git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@12589 a8845c50-7012-0410-95d3-8e1449b9b1e4