This is the first set of changes needed to implement PPI functionality required for release 1.1 -- see http://mingle.mifos.org:7070/projects/mifos/cards/443 for more information.
This set of changes defines new a new business object -- PpiLikelihoodChart -- to hold the likelihood table associated with a PPI survey, as well as supporting classes and new unit test classes.

The business objects do not yet have Hibernate mappings or associated database table definitions -- to be completed in a future revision. Also needed is the ability to parse the likelihood chart from XML.

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@12659 a8845c50-7012-0410-95d3-8e1449b9b1e4