Covered the fact thoroughly that we're now reusing a permission/activity/whateveryoucallit: SecurityConstants#CAN_VIEW_ADMIN_DOCUMENTS. The two things this permission is used for are: * listing all available admin documents. From the UI: "Admin->View admin   documents" * running/executing an admin document when it shows up during a   particular loan or savings account state
This was done to fix the issue quickly and since the operations seem to have a reasonable amount of overlap.

A specific unit test for this issue isn't necessary since it uses one of the standard methods for permissions (the other being per-report permissions). The permissions system used for reports seems bolted-on (see MifosRequestProcessor#isReportRequest) and will likely need refactoring.

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@13163 a8845c50-7012-0410-95d3-8e1449b9b1e4