Fix for issue 2017 (https://mifos.dev.java.net/issues/show_bug.cgi?id=2017)
I found two bugs which I've fixed:
 - MeetingBO.getFirstDateForWeek(startdate), whereby if for example the
startdate was Sunday 8 June and the current meeting date was a Saturday, the
date returned was Saturday 7 June, when you'd expect it to return Saturday 14 June.
 - MeetingBO.isValidMeetingDate(), where this method didn't take into
consideration when the isRepaymentIndepOfMeetingEnabled flag is true.

I've also removed DateUtils.getFirstDateForDayOfWeekAfterDate() (and relevant unit tests), which was added by Keith as an attempt to refactor MeetingBO.getFirstDateForWeek(), but since the fix touches this method, I've undone the refactoring for the time being so as to make sure no further bugs are introduced when we're so close to the 1.1 release.


git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@13267 a8845c50-7012-0410-95d3-8e1449b9b1e4