Renabling 4 tests in TestLoanBORedoDisbursal commented out in r13264.
These tests depended on not running on a Monday in order to pass (due to low level constructor defaults).  The dependence on Monday has been removed.  Eventually we would like to have tests that run on a fixed date, but there are currently too many test harness methods that depend on using today's date to do this without major refactoring.

A utility method and tests were added to DateUtils for mapping a java.util.Date to a mifos WeekDay.


git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@13275 a8845c50-7012-0410-95d3-8e1449b9b1e4