Fix the fix for issue 2081.
After logging out, the session is gone, so there is no user context and no preferred locale available. Hopefully the one-argument form of ResourceBundle.getBundle() is a better option if the session is empty.

This was breaking acceptance tests that, for instance, start out with "GET /mifos/loginAction.do?method=logout". Had we automated acceptance tests, we would have caught this regression sooner.

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@14198 a8845c50-7012-0410-95d3-8e1449b9b1e4