Fix for issue 2237 (https://mifos.dev.java.net/issues/show_bug.cgi?id=2237)
Comments from contributor Viswanath Durbha:

yoursettings.jsp - 

A presence of the character '#' in the userName query parameter causes problem as '#' has special meaning in URLs. So it had to be encode with %23. So the html tag <a> is replaced with the tag <c:url> and <c:param> which automatically encode the userName query parameter.



git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@14234 a8845c50-7012-0410-95d3-8e1449b9b1e4