[FIXED MIFOS-1567] Make timer thread created by MifosScheduler a daemon thread. This allows Tomcat to shut down cleanly when Mifos is deployed.
Mifos was preventing clean shutdown of the Tomcat servlet container. This was
found using the following techniques:
* running the Mifos Web application in the Eclipse WTP in debug mode
  * this exposed the fact that a non-daemon thread was present when a
    Tomcat shutdown was requested
  * stepping through ApplicationInitializer confirmed the misbehaving
    non-daemon thread was created by MifosScheduler

MifosXMLWatchdog, the MySQL driver, and c3p0 also create threads, but these are
all daemon threads, hence are subject to immediate cancellation when the
servlet context is destroyed.

Another technique that was useful was turning on (or turning up?) c3p0 logging. This was done by creating c3p0.properties with the following settings:

  com.mchange.v2.log.MLog=com.mchange.v2.log.FallbackMLog

and placing this file on the application server classpath.

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@14864 a8845c50-7012-0410-95d3-8e1449b9b1e4