Re-did much of the select box selections in the page objects. From choosing on label to choosing on the value of the choice. This is done because the labels change depending on the locale but the value stays the same.
This change involves not using the literal value that's shown in the box ("Cash") as an argument but a constant, SomeParameterClass.CASH, as the argument for "set<Whatever>".

An update to the wiki page about writing acceptance tests will follow shortly.

This change also reduces the number of tests that fail in French to ~20 (down from 70 at the beginning of the week).

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@15482 a8845c50-7012-0410-95d3-8e1449b9b1e4