Changed the error message in bulkentry.jsp to be inside of span tags, like all the other error messages in other jsp's. Also fixed a broken tag to make sure that a descriptive text was displayed (no idea how long this has been broken for).
Added a lot of javadoc to the helper classes.

Added a method to MifosPage to verify that an error message is displayed.
Then changed a couple of tests from verifying that a specific (english) error message is displayed to just verifying that an error message is displayed at all. This was made to ensure that the acceptance tests don't fail in other locales.

There are now just 2 tests failing when running the acceptance tests in french :)

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@15527 a8845c50-7012-0410-95d3-8e1449b9b1e4