Udai Gupta wrote: > I was getting this failure on my system. I was just wondering is this > a system specific failure. Can anyone with windows try this test and see > if they are also getting this failure or its just me. > > java.lang.AssertionError: correct size Expected: is <12> got: <13> >   at org.junit.Assert.assertThat(Assert.java:502) >   at org.mifos.framework.util.UnicodeUtilTest.canDetectUtf8(UnicodeUtilTest.java:92)
Ah, I bet that's because the file ended up with CRLF line endings when
checked out on Windows. I set svn:eol-style to force UNIX line endings
so file is 12 bytes long. If this doesn't work, we can just comment out
the assertion for bytes read, it isn't really a useful test, anyway. The
important part is verifying the contents of the file.


git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16027 a8845c50-7012-0410-95d3-8e1449b9b1e4