Resolving issue 2517, card 2271 (localizations not being included in war)
The problem was that the generation of localized properties files (in the integration test phase) was occurring after the WAR file had been built (in the package phase).

In maven plugins within the same phase are supposed to execute in the order defined in the pom, but according to an outstanding maven bug (http://jira.codehaus.org/browse/MNG-2258) they don't. So just changing this plugin to run in the package phase and moving it before the war plugin didn't resolve the issue.

By setting the phase to compile (which occurs before the package phase) then the properties files are generated before the war is and so they are included as expected.


git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16169 a8845c50-7012-0410-95d3-8e1449b9b1e4