This resolves issues 2610 and 2619
For 2610, "Collection Sheet Entry exception when required fields missing on select form ", a few acceptance tests have been added

Although 2619, "Error message when overpaying loan in CSE doesn't display full loan account ID" was stated as a Gazelle C fix, it was easier to change it than not to.  Manual check to verify.



git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16356 a8845c50-7012-0410-95d3-8e1449b9b1e4