Save Collection Sheet performance tuning (where it reads the existing data to build up the object model prior to persisting).
Concerns accountPaymentEntity/account_payment.  Also affects non collection sheet transactions that create account_payments.

when processing new transactions for accounts (e.g. repayments, disbursals) account_payments (table entries) are added.  There is no need to refer to or process previous account_payments.  However, all previous account_payments for an account were retrieved.  For loans with large numbers of installments this was quite an overhead.

According to my debugging, the cause of this is that hibernate is configured to retrieve account_payments using the <set></set> mapping (which seems to be the standard way of mapping one to many's).  I suspect that in order to guarantee set uniqueness hibernate retrieves existing account_payments to check that the ones being added are unique (which they always are and would get caught by primary key violation if they weren't).

When I changed accountBo.hbm.xml mapping for accountPayments to a <bag></bag> mapping this retrieving of existing account payments when processing an account did not occur.

It did involve a number of changes to make accountPayments a List rather than a Set.  However, much of the code I changed wanted to process accountPayments as a list anyway which made me thing there might be a history to this (any idea Devs?)  Also, in many cases, the application expects accountPayments in descending paymentId order.  <bag> allows that to happen in the same way <set> did (but <list> doesn't)... maybe it something to do with that?  

Also some minor tidy ups.



git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16550 a8845c50-7012-0410-95d3-8e1449b9b1e4