This commit contains a change relating to the batch job GenerateMeetingsForCustomerAndSavings.
The change is to only return customer and savings accounts that need to have more schedules added.

This cuts out a lot of fixed data retrieval. i.e if there are 346k customer_accounts having 43million customer_schedules then previously all (if active) were read in and processed giving a high fixed cost even if no schedules needed to be added.

The other 'variable' side to this batch job is in generating the extra installments.  This is extremely costly due to an explosion of holiday queries when adding new installments (Prabaharan's analysis is spot on - http://groups.google.com/group/mifosdeveloper/msg/934d7f3c0b5613dc ).  The more schedules that need to be generated the more important it becomes and it is this variability that risks batch jobs over-running into next days operational work. i.e. GenerateMeetingsForCustomerAndSavings is by far and away the riskiest job.

Will ask about status of Prabaharan's fix later on in the mailing list.

With this fix and Prabaharan's fix I'd expect the GK GenerateMeetingsForCustomerAndSavings to be no more than 15 or 30 mins in length (currently varies between 2 and 8 hours) which would mean scalability emphasis could be taken off the batch jobs until some of the work Keith W is doing to remove customer schedules is implemented.


git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16708 a8845c50-7012-0410-95d3-8e1449b9b1e4