FIXED MIFOS-2709, Have to remove .0 in periodic loan fee when attaching to loan account in loan account creation with digitsafterdecimal = 0
Additional fees validation fixed for Rate Fee and Amount Fee.

Only Rate fee has trailing zero.
Now the validation will not fail because rate fee uses digitsAfterDecimalForInterest but the trailing zero will not be removed because of double to string conversion. To get right behavior i.e not having trailing zero for rate fees at digitsAfterDecimal=0 we need to get rid of double data type for Rate and use BigDecimal and add formatting like Money for BigDecimal Rafe fees. The change for Double to BigDecimal data type for rate fee has not been made in this fix, because it will hit lots of places in code and since the trunk has lots of change compared to 1.4.x branch it will be difficult to merge those changes for hotfix. 

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16736 a8845c50-7012-0410-95d3-8e1449b9b1e4