More on Migration from hbm.xml to JPA 1.0 (Hibernate JPA 1.0 implementation. More details http://mifosforge.jira.com/browse/MIFOS-2696.
Removed digitsAfterDecimal from MifosCurrency constructor, because this field has been deleted in previous database upgrades.
Made roundingAmount to be mapped with BigDecimal instead of Float, because it is being converted as BigDecimal for use in many places and BigDecimal type will have no effect in code except now instead of creating BigDecimal from Float the field will already be mapped as BigDecimal.



git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16755 a8845c50-7012-0410-95d3-8e1449b9b1e4