adding 'public' to cache-control response headers in the belief that it helps browser caching (at least for firefox) if https/SSL is used.
http://code.google.com/speed/page-speed/docs/caching.html#LeverageBrowserCaching

"Use the Cache control: public directive to enable HTTPS caching for Firefox.
Some versions of Firefox require that the Cache control: public header to be set in order for resources sent over SSL to be cached on disk, even if the other caching headers are explicitly set. Although this header is normally used to enable caching by proxy servers (as described below), proxies cannot cache any content sent over HTTPS, so it is always safe to set this header for HTTPS resources."

git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@17053 a8845c50-7012-0410-95d3-8e1449b9b1e4