[FIXED MIFOS-2871] Upgrade error on upgrade to version 232 with GK data.
n upgrade 227 a column in Fee table with FK to Currency was added and upgrade 232 is removing that column because later we decided not to have that column.

The fix is straight forward, just deleted the addition and deletion of that column from respective upgrades as they are both in same release.


git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@17190 a8845c50-7012-0410-95d3-8e1449b9b1e4