LoanDisbursementAction → use disburseLoans() API
Both LoanDisbursementAction (UI layer) and AccountService (API)
disburse loans, so this allows them to share code. This is not ideal
since code internal to Mifos (LoanDisbursementAction) is forced
through an *external* API (AccountService). Potential issues: (a) it
may be more convenient/sensical to use more Mifos objects when
disbursing loans and (b) to receive more sensical/internal
exceptions from AccountService.

Van and I paired on this today, and Van's guidance was to provide
disburseLoans() in LoanService, moving towards a call path like the
following for loan disbursals from the two entry points:

1. LoanDisbursementAction → LoanServiceFacadeWebTier → LoanService
2. AccountService → LoanService

This work is related to Mingle story #2441.