FIXED MIFOS-3063 Remove DatabaseVersionPersistence class
Remove DatabaseVersionPersistence class and resloved compilation issues with the numerous touch-points of this class.
The touch-points that were not non-trivial have been marked with the 'TODO Update for NSDU' comment.

Fixed broken build by getting rid of DatabaseMigrator dependancies on classes in the Test module.
Some of the tests that lived DatabaseVersionPersistenceIntegrationTest were also found to be good test cases for
DatabaseMigrator as well. Some of the test cases were also found to be belong in SqlExecutorIntegrationTest.
Dev-Task MIFOS-3206 will take care of that.

The testMethodUpgrade method has also been disabled and will be reimplemented later.