Share ApplicationContext between dataset upgrades.
Each time Spring is initialized, new database connections are created.
For 68 datasets, this was exhausting the 100 or so default connections
on my MySQL server. After a quick chat with Jakub, Kojo, Van and myself
on IRC, we decided that the Spring ApplicationContext could be shared
between dataset upgrades. Not only does this use a fixed number of
database connections, it runs significantly faster (since initializing
the Spring configuration is expensive).