FIXED MIFOS-4111 when activating a question group, different questions with the same question text are not imported as unique questions.
When creating questions, if a question has no nickname (nickname == null),
then try to find an existing question with the same text and use it
instead.  If a question we're about to create has a unique nickname,
then go ahead and create a new question even if there is an existing
question with the same text.  If there is an an existing question with
a nickname that matches the new question nickname, then use the existing
question instead.