FIXED MIFOS-1182: second attempt at fixing Firefox back button issue.
The previous fix no longer works in Firefox. Part of the reason is
when user clicks the back button, a cached version of the page along
with the state it was in gets displayed. Because of this, the document
ready/onload event does not get fired.

The new fix involves binding the button re-enable logic to "pageshow" event
rather than body onload. See this page for me information:
https://developer.mozilla.org/En/Using_Firefox_1.5_caching.