According to JPA 2 : @Id is already on a field which make AccessType.FIELD redundant
you can define the @AccessType annotaiton on, an entity, a superclass, an embeddable object, a property
The access type is overriden for the annotated element,
if overriden on a class, all the properties of the given class inherit the access type.
For root entities, the access type is considered to be the default one for other elements.
If the access type is marked as "property", the getters are scanned for annotations,
if the access type is marked as "field", the fields are scanned for annotations.
Otherwise the elements marked with @Id or @embeddedId are scanned