Fix errexit operation for generic Hudson deploy script.
The short-circuit didn't work since it was *not* setting
can_hit_test_server in the case where curl actually succeeded!

It would be fine to reverse the boolean, e.g.:

  cannot_hit_test_server=0
  while [ can_hit_test_server -eq 0 ]
  do
      something
      can_hit_test_server=$?
  done

But I think a clearer path is to simply turn off errexit
(using "set +e") when we don't want it on.