Cache CustomizedText in hibernate 2nd-level cache.
While investigating a performance issue, we've noticed that there are no
cached entities or queries in Mifos, even though many cacheable database
lookups are performed with every page request. This doesn't seem to
become a problem unless the database is on a separate machine. This is
all educated guesses, I have no data to back this up.

This is a start of caching some database interaction.

This enables caching only for CustomizedText entities, which means this
only provides a benefit if replacements (rows in customized_text) are
present. A subsequent change will be more generally useful, caching a
very often-run "select * from customized_text..." query.

Notice the change to *hibernate.cfg.xml to set the cache provider... I
was unable to turn on caching by modifying dataSourceContext.xml
instead. This may be because there are two copies of
dataSourceContext.xml, or I was making some other error.

I picked NONSTRICT_READ_WRITE as the strategy for caching the
CustomizedText entity because replacements should rarely be updated (how
often would you expect to change how the word "Clients" appears?), and
the worst that should happen is that someone wouldn't see a replacement
the instant it was created/edited and would have to reload a page.

Since I chose ehcache as the cache provider, I added back the phone home
off switch in ApplicationInitializer, originally added in
c96cb21d948ed7d7206094ee43efd0b73671fd9b, then removed in
b2da574bb528daddc793ec6171b03bfbcbe1e72a.

We may not want to use the 2nd-level cache at all in tests, but until we
know that for sure I enabled it across the board to keep our testing
environment as similar as possible to runtime.

See also:
http://thread.gmane.org/gmane.comp.finance.mifos.devel/11469