Enable query caching in the hibernate 2nd-level cache.
This will be utilized in a subsequent commit for caching "select * from
customized_text...".

We may not want to use the 2nd-level cache at all in tests, but until we
know that for sure I enabled it across the board to keep our testing
environment as similar as possible to runtime.