FIXED MIFOS-4861 - made Add/Remove buttons translatable
I chose to manually concatenate "Add" + " >>" because I feel like that
layout (placing >> to the right of the verb Add) should be fixed (ie:
unmodifiable by a translation). Same for "<< Remove".

I originally tried using LabelTagUtils to lookup "add" and "remove"
messages, but fixing MifosSelectTest was difficult because I had to
initialize too much Struts infrastructure to make the lookup happen. Van
had the awesome idea to instead use MessageLookup, and make it
injectable in MifosSelect.

We first made sure that changing the implementation in MifosSelect to
use MessageLookup worked, then we mocked MessageLookup in
MifosSelectTest.

MessageLookup is legacy code, but so is Struts/JSP, so all this should
go away once the migration to SpringMVC/FTL is complete.

I converted MifosSelectTest to JUnit 4 using AccountingDaoTest as an
example. The MockitoJUnitRunner wasn't working with JUnit 3.

After making the "Add" button translatable, I notice that it's probably
too narrow for longer translations, for instance, German.