MIFOS-4950.  Removed part of Upgrade1283765911 that migrated historic batch job records.
This was the only place in the code that relied on SessionFactoryUtils.getDataSource.  After refactoring database connection configuration, this call no longer worked.  This means that this upgrade no longer copied legacy batch job records to the new batch job history table.  If a user wants to preserve these records, they would need to upgrade to Mifos 2.0.2 first, then move to Mifos 2.0.3.  Migrating directly to 2.0.3 will work, but batch job records will not be migrated.