MIFOS-4965.  Changed call to flush and clear session to just flush the session.
In CustomerServiceImpl.transferGroupTo(GroupBO group, OfficeBO transferToOffice), data loaded into the session is needed to complete the transfer so it shouldn't be cleared.  There may be other calls where only a flush should be done. Originally the HibernateTransactionHelp.flushSession() call was doing a flushAndClearSession on StaticHibernateUtils.  It's not clear if this is what was intended or if it is needed.  At least in most usage, it should probably be just a call to flush.
(cherry picked from commit a44ce1bf3a8ef9edfc8b602c72930b486de3d431)